Cleanup - unify getter names and helper files
